+++ examples/resources/metabase_database/resource.tf
@@ -1,0 +1,8 @@
+resource "metabase_database" "example" {
+  engine = "h2"
+  name   = "H2 database"
+
+  details = jsonencode({
+    db = "zip:/app/metabase.jar!/sample-database.db;USER=GUEST;PASSWORD=guest"
+  })
+}